made listingsData its own file to call for all parts of website
