+++ client/src/components/pages/AllProducts.js
@@ -1,0 +1,14 @@
+import Owl from './owl.png';
+import test from './test.png';
+
+const listingsData = [
+  { id: 1, title: 'Item Name', seller: "aaron", description: "very good", price: '$100.00', image: test, pickup: 'on-campus', category: 'insert tags'},
+  { id: 2, title: 'Item saf 2', seller: 'aaron', description: "worth an investment", price: '$50.50', image: Owl, pickup: 'off-campus', category: 'insert tags' },
+  { id: 3, title: 'dsfa 2', seller: 'aaron', description: "worth an investment", price: '$50.50', image: Owl, pickup: 'off-campus', category: 'insert tags' },
+  { id: 4, title: 'bdfb Name 2', seller: 'aaron', description: "worth an investment", price: '$50.50', image: test, pickup: 'off-campus', category: 'insert tags' },
+  { id: 5, title: '3242 Name 2', seller: 'dfsdfsd sdf', description: "worth an investment", price: '$50.50', image: Owl, pickup: 'off-campus', category: 'insert tags' },
+  { id: 6, title: '543 2', seller: 'dfgfd Doe', description: "worth an investment", price: '$50.50', image: test, pickup: 'off-campus', category: 'insert tags' },
+];
+
+export default listingsData;
+
