+++ client/src/components/pages/HomePage.js
@@ -20,21 +20,11 @@ import {
 import { Navbar } from "./Navbar";
 
 import Button from '@mui/material/Button';
-import Owl from './owl.png';
-import test from './test.png';
 import ExpandMoreIcon from '@mui/icons-material/ExpandMore';
-
 
 import favoritesIcon from "../../assets/favorites-icon.svg";
 
-const listingsData = [
-  { id: 1, title: 'Item Name', seller: "aaron", description: "very good", price: '$100.00', image: Owl, pickup: 'on-campus', category: 'insert tags'},
-  { id: 2, title: 'Item saf 2', seller: 'dsf', description: "worth an investment", price: '$50.50', image: test, pickup: 'off-campus', category: 'insert tags' },
-  { id: 3, title: 'dsfa 2', seller: 'sdf', description: "worth an investment", price: '$50.50', image: test, pickup: 'off-campus', category: 'insert tags' },
-  { id: 4, title: 'bdfb Name 2', seller: 'John sdf', description: "worth an investment", price: '$50.50', image: test, pickup: 'off-campus', category: 'insert tags' },
-  { id: 5, title: '3242 Name 2', seller: 'dfsdfsd sdf', description: "worth an investment", price: '$50.50', image: test, pickup: 'off-campus', category: 'insert tags' },
-  { id: 6, title: '543 2', seller: 'dfgfd Doe', description: "worth an investment", price: '$50.50', image: test, pickup: 'off-campus', category: 'insert tags' },
-];
+import listingsData from './AllProducts.js';
 
 const CardComponent = ({ id, image, title, price, seller }) => {
   return (
